+++ mylar/helpers.py
@@ -48,9 +48,7 @@ from lib.rarfile import rarfile
 
 import mylar
 from . import logger
-from mylar import db, sabnzbd, nzbget, process, getcomics, getimage
-from mylar.downloaders import mega, pixeldrain, mediafire
-from mylar.downloaders.jdownloader2 import JDownloader2
+from mylar import db, getimage
 
 
 def multikeysort(items, columns):
@@ -2888,636 +2886,6 @@ def latestissue_update():
             except Exception as e:
                 logger.fdebug('exception encountered: %s' % e)
                 continue
-
-def ddl_downloader(queue):
-    myDB = db.DBConnection()
-    link_type_failure = {}
-    while True:
-        if mylar.DDL_LOCK is True:
-            time.sleep(5)
-
-        elif mylar.DDL_LOCK is False and queue.qsize() >= 1:
-            item = queue.get(True)
-
-            if item == 'exit':
-                logger.info('Cleaning up workers for shutdown')
-                break
-
-            if item['id'] not in mylar.DDL_QUEUED:
-                mylar.DDL_QUEUED.append(item['id'])
-
-            try:
-                link_type_failure[item['id']].append(item['link_type_failure'])
-            except Exception:
-                pass
-
-            #logger.info('[%s] link_type_failure: %s' % (item['id'], link_type_failure))
-
-            logger.info('Now loading request from DDL queue: %s' % item['series'])
-
-            #write this to the table so we have a record of what's going on.
-            ctrlval = {'id':      item['id']}
-            val = {'status':       'Downloading',
-                   'updated_date': datetime.datetime.now().strftime('%Y-%m-%d %H:%M')}
-            myDB.upsert('ddl_info', val, ctrlval)
-
-            if item['site'] == 'DDL(GetComics)':
-                try:
-                    remote_filesize = item['remote_filesize']
-                except Exception:
-                    try:
-                        remote_filesize = helpers.human2bytes(re.sub('/s', '', item['size'][:-1]).strip())
-                    except Exception:
-                        remote_filesize = 0
-
-                if any([item['link_type'] == 'GC-Main', item['link_type'] == 'GC_Mirror']):
-                    ddz = getcomics.GC()
-                    ddzstat = ddz.downloadit(item['id'], item['link'], item['mainlink'], item['resume'], item['issueid'], remote_filesize)
-                elif item['link_type'] == 'GC-Mega':
-                    meganz = mega.MegaNZ()
-                    ddzstat = meganz.ddl_download(item['link'], None, item['id'], item['issueid'], item['link_type']) #item['filename'], item['id'])
-                elif item['link_type'] == 'GC-Media':
-                    mediaf = mediafire.MediaFire()
-                    ddzstat = mediaf.ddl_download(item['link'], item['id'], item['issueid']) #item['filename'], item['id'])
-                elif item['link_type'] == 'GC-Pixel':
-                    pdrain = pixeldrain.PixelDrain()
-                    ddzstat = pdrain.ddl_download(item['link'], item['id'], item['issueid']) #item['filename'], item['id'])
-
-            elif item['site'] == 'DDL(External)':
-                meganz = mega.MegaNZ()
-                ddzstat = meganz.ddl_download(item['link'], item['filename'], item['id'], item['issueid'], item['link_type'])
-
-            # Check for file validity post download and mark as failure if file is not a zip, rar, or pdf
-            # Can only check single downloads.  Packs will have to be managed by post-processing if enabled
-            if ddzstat['success'] and ddzstat['filename'] is not None:
-                filecondition = check_file_condition(ddzstat['path'])
-                if not filecondition['status']:
-                    logger.warn(f"CRC Check: File {ddzstat['path']} failed condition check ({filecondition['quality']}).  Marking as failed.")
-                    ddzstat['success'] = False
-                    ddzstat['link_type_failure'] = item['link_type']
-
-            if ddzstat['success'] is True:
-                tdnow = datetime.datetime.now()
-                nval = {'status':  'Completed',
-                        'updated_date': tdnow.strftime('%Y-%m-%d %H:%M')}
-                myDB.upsert('ddl_info', nval, ctrlval)
-
-            if all([ddzstat['success'] is True, mylar.CONFIG.POST_PROCESSING is True]):
-                try:
-                    if ddzstat['filename'] is None:
-                        logger.info('%s successfully downloaded - now initiating post-processing for %s.' % (os.path.basename(ddzstat['path']), ddzstat['path']))
-                        mylar.PP_QUEUE.put({'nzb_name':     os.path.basename(ddzstat['path']),
-                                            'nzb_folder':   ddzstat['path'],
-                                            'failed':       False,
-                                            'issueid':      None,
-                                            'comicid':      item['comicid'],
-                                            'apicall':      True,
-                                            'ddl':          True,
-                                            'download_info': {'provider': 'DDL', 'id': item['id']}})
-                    else:
-                        logger.info('%s successfully downloaded - now initiating post-processing for %s' % (ddzstat['filename'], ddzstat['path']))
-                        mylar.PP_QUEUE.put({'nzb_name':     ddzstat['filename'],
-                                            'nzb_folder':   ddzstat['path'],
-                                            'failed':       False,
-                                            'issueid':      item['issueid'],
-                                            'comicid':      item['comicid'],
-                                            'apicall':      True,
-                                            'ddl':          True,
-                                            'download_info': {'provider': 'DDL', 'id': item['id']}})
-                except Exception as e:
-                    logger.error('process error: %s [%s]' %(e, ddzstat))
-
-                #logger.fdebug('mylar.ddl_queued: %s' % mylar.DDL_QUEUED)
-                mylar.DDL_QUEUED.remove(item['id'])
-                try:
-                    link_type_failure.pop(item['id'])
-                except KeyError:
-                    pass
-
-                try:
-                    pck_cnt = 0
-                    if item['comicinfo'][0]['pack'] is True:
-                        logger.fdebug('[PACK DETECTION] Attempting to remove issueids from the pack dont-queue list')
-                        for x,y in dict(mylar.PACK_ISSUEIDS_DONT_QUEUE).items():
-                            if y == item['id']:
-                                pck_cnt +=1
-                                del mylar.PACK_ISSUEIDS_DONT_QUEUE[x]
-                        logger.fdebug('Successfully removed %s issueids from pack queue list as download is completed.' % pck_cnt)
-                except Exception:
-                    pass
-
-                # remove html file from cache if it's successful
-                ddl_cleanup(item['id'])
-
-            elif all([ddzstat['success'] is True, mylar.CONFIG.POST_PROCESSING is False]):
-                path = ddzstat['path']
-                if ddzstat['filename'] is not None:
-                    path = os.path.join(path, ddzstat['filename'])
-                logger.info('File successfully downloaded. Post Processing is not enabled - item retained here: %s' % (path,))
-                ddl_cleanup(item['id'])
-            else:
-                if item['site'] == 'DDL(GetComics)':
-                    try:
-                        ltf = ddzstat['links_exhausted']
-                    except KeyError:
-                        logger.info('[Status: %s] Failed to download item from %s : %s ' % (ddzstat['success'], item['link_type'], ddzstat))
-                        try:
-                            link_type_failure[item['id']].append(item['link_type'])
-                        except KeyError:
-                            link_type_failure[item['id']] = [item['link_type']]
-                        logger.fdebug('[%s] link_type_failure: %s' % (item['id'], link_type_failure))
-                        ggc = getcomics.GC(comicid=item['comicid'], issueid=item['issueid'], oneoff=item['oneoff'])
-                        ggc.parse_downloadresults(item['id'], item['mainlink'], item['comicinfo'], item['packinfo'], link_type_failure[item['id']])
-                    else:
-                        logger.info('[REDO] Exhausted all available links [%s] for issueid %s and was not able to download anything' % (link_type_failure[item['id']], item['issueid']))
-                        nval = {'status':  'Failed',
-                                'updated_date': datetime.datetime.now().strftime('%Y-%m-%d %H:%M')}
-                        myDB.upsert('ddl_info', nval, ctrlval)
-                        #undo all snatched items, to previous status via item['id'] - this will be set to Skipped currently regardless of previous status
-                        reverse_the_pack_snatch(item['id'], item['comicid'])
-                        link_type_failure.pop(item['id'])
-                        ddl_cleanup(item['id'])
-                else:
-                    logger.info('[Status: %s] Failed to download item from %s : %s ' % (ddzstat['success'], item['site'], ddzstat))
-                    myDB.action('DELETE FROM ddl_info where id=?', [item['id']])
-                    mylar.search.FailedMark(item['issueid'], item['comicid'], item['id'], ddzstat['filename'], item['site'])
-        else:
-            time.sleep(5)
-
-def ddl_cleanup(record_id):
-    if getattr(mylar.CONFIG, 'KEEP_HTML_CACHE', False):
-        logger.fdebug('[HTML-cleanup] KEEP_HTML_CACHE enabled; skipping removal for %s.', record_id)
-        return
-
-    tlnk = 'getcomics-%s.html' % record_id
-    cache_path = os.path.join(mylar.CONFIG.CACHE_DIR, 'html_cache', tlnk)
-    try:
-        os.remove(cache_path)
-    except FileNotFoundError:
-        logger.fdebug('[HTML-cleanup] %s not found in html_cache. Nothing to remove.', tlnk)
-    except Exception as e:
-        logger.fdebug('[HTML-cleanup] Unable to remove %s from html_cache: %s. '
-                    'Manual removal required or set `cleanup_cache=True` in the config.ini to '
-                    'clean cache items on every startup. If this was a Retry - ignore this.',
-                    tlnk, e)
-
-
-def jd2_queue_monitor(queue):
-    jd2_client = JDownloader2()
-    myDB = db.DBConnection()
-    while True:
-        if queue.qsize() >= 1:
-            item = queue.get(True)
-            if item == 'exit':
-                logger.info('[JD2-QUEUE] Cleaning up workers for shutdown')
-                break
-            if item == 'startup':
-                try:
-                    pending_jd2 = myDB.select(
-                        "SELECT * FROM ddl_info WHERE jd2_job_id IS NOT NULL AND status = 'Queued'"
-                    )
-
-                except Exception as err:
-                    logger.warn('[JD2-QUEUE] Unable to repopulate JD2 monitor queue: %s', err)
-                else:
-                    logger.info('[JD2-QUEUE] Repopulating JD2 monitor queue')
-                    for job in pending_jd2 or []:
-                        job_data = dict(job)
-                        job_id = job_data.get('jd2_job_id')
-                        record_id = job_data.get('ID') or job_data.get('id')
-                        series = job_data.get('series')
-                        year = job_data.get('year')
-                        filename = job_data.get('tmp_filename') or job_data.get('filename')
-                        if not filename:
-                            if series and year:
-                                filename = f"{series} ({year}) - {record_id}"
-                            elif series:
-                                filename = series
-                            elif record_id:
-                                filename = str(record_id)
-                        queue_payload = {
-                            'link': job_data.get('link'),
-                            'mainlink': job_data.get('mainlink'),
-                            'series': series,
-                            'year': year,
-                            'size': job_data.get('size'),
-                            'comicid': job_data.get('comicid'),
-                            'issueid': job_data.get('issueid'),
-                            'oneoff': job_data.get('oneoff'),
-                            'id': record_id,
-                            'link_type': job_data.get('link_type'),
-                            'filename': filename,
-                            'comicinfo': job_data.get('comicinfo'),
-                            'packinfo': job_data.get('packinfo'),
-                            'site': job_data.get('site') or 'DDL(GetComics)',
-                            'remote_filesize': job_data.get('remote_filesize') or 0,
-                            'resume': None,
-                            'jd2_job_id': job_id,
-                        }
-                        queue.put(queue_payload)
-                continue
-            # if the job id is 0 it's a special case where we need to run the download with self.jd2.submit(jd2_priority_links, package_name, record_id=mod_id). 
-            
-            
-            if item.get('jd2_job_id') == 0:
-                jd2_priority_links = item.get('jd2_priority_links')
-                queue_payload = dict(item)
-                queue_payload.pop('jd2_job_id', None)
-                queue_payload.pop('jd2_priority_links', None)
-                if not jd2_priority_links:
-                    logger.warn('Priority links don\'t exist; cannot process JD2 job. Falling back to DDL queue.')
-                    mylar.DDL_QUEUE.put(queue_payload)
-                    continue
-                record_id = item.get('id') or item.get('ID')
-                package_name = '%s (%s) - %s' % (item.get('series'), item.get('year'), record_id)
-                submit_result = jd2_client.submit(jd2_priority_links, package_name, record_id)
-                
-                if not submit_result.get('status'):
-                    logger.warn('[JD2] Submission failed for %s; falling back to DDL queue. Details: %s', package_name, submit_result)
-                    mylar.DDL_QUEUE.put(queue_payload)
-                else:
-                    logger.debug('[JD2] Submission response for %s: %s', package_name, submit_result)
-                    job_id = submit_result.get('jobid')
-                    if not job_id:
-                        mylar.DDL_QUEUE.put(queue_payload)
-                        continue
-                    queue_payload['jd2_job_id'] = job_id
-                    mylar.JD2_QUEUE.put(queue_payload)
-                    logger.info('[JD2] JD2 job %s submitted for %s', job_id, package_name)
-                    time.sleep(10) # allow the job to start
-                    continue
-                    
-            logger.info('[JD2-QUEUE] Now loading from queue: %s' % item)
-            record_id = item.get('ID') or item.get('id')
-            job_id = item.get('jd2_job_id') if isinstance(item, dict) else None
-            
-
-            if jd2_client is not None and job_id:
-                try:
-                    status_payload = jd2_client.status(job_id)
-                except Exception as err:
-                    logger.warn('[JD2-QUEUE] Error polling job %s: %s', job_id, err)
-                else:
-                    logger.info('[JD2-QUEUE] status update for job %s (record %s): %s', job_id, record_id, status_payload)
-                    job_status = (status_payload or {}).get('status')
-                    job_found = (status_payload or {}).get('found')
-                    
-                    completed_states = {'FINISHED', 'Finished', 'finished', 'DONE', 'Done', 'Extraction OK', '[SHA256] CRC OK', 'Finished(Mirror)'}
-                    failed_states = {'ERROR', 'FAILED', 'Error', 'Failed'}
-
-                    if not job_found:
-                        stale = False
-                        try:
-                            row = myDB.selectone("SELECT updated_date FROM ddl_info WHERE id=?", [record_id]).fetchone()
-                        except Exception as err:
-                            logger.warning('JD2 stale check failed for record %s: %s', record_id, err)
-                            updated_str = None
-                        else:
-                            if row is None:
-                                logger.info('[JD2-QUEUE] No ddl_info entry found for id %s; dropping job %s from monitoring.', record_id, job_id)
-                                continue
-                            updated_str = row['updated_date']
-
-                        if updated_str:
-                            try:
-                                updated_dt = datetime.datetime.strptime(updated_str, '%Y-%m-%d %H:%M')
-                            except Exception:
-                                stale = False
-                            else:
-                                stale = (datetime.datetime.now() - updated_dt) >= datetime.timedelta(minutes=3)
-
-                        if stale:
-                            if myDB is not None:
-                                myDB.upsert(
-                                    'ddl_info',
-                                    {
-                                        'status': 'Failed',
-                                        'updated_date': datetime.datetime.now().strftime('%Y-%m-%d %H:%M'),
-                                    },
-                                    {'id': record_id},
-                                )
-                            continue
-
-                        queue.put(item)
-                        time.sleep(10)
-                        continue
-                    
-                    if job_status in completed_states:
-                        myDB.upsert(
-                                'ddl_info',
-                                {
-                                    'status': 'Completed',
-                                    'updated_date': datetime.datetime.now().strftime('%Y-%m-%d %H:%M'),
-                                },
-                                {'id': record_id},
-                            )
-                        logger.info('[JD2-QUEUE] Download for %s completed; status marked as Completed.', record_id)
-                        
-                        if not isinstance(item, dict):
-                            logger.warn('[JD2-QUEUE] Unexpected item type %s for job %s; skipping post-processing.', type(item), job_id)
-                            continue
-                        job_filename = (status_payload or {}).get('data').get('name')
-                        if mylar.CONFIG.POST_PROCESSING is True:
-                            dest_root = getattr(mylar.CONFIG, 'JD2_DEST_DIR', None) or None
-                            folder_name = (
-                                "%s - %s" % (item.get('filename'), record_id)
-                                if item.get('filename')
-                                else "%s (%s) - %s" % (series, year, record_id)
-                            )
-                            nzb_folder = os.path.join(dest_root, folder_name)
-                            
-                            
-                            if nzb_folder:
-                                try:
-                                    mylar.PP_QUEUE.put({
-                                        'nzb_name': job_filename or 'Manual Run',
-                                        'nzb_folder': nzb_folder,
-                                        'failed': False,
-                                        'issueid': item.get('issueid'),
-                                        'comicid': item.get('comicid'),
-                                        'apicall': True,
-                                        'ddl': True,
-                                        'download_info': {'provider': 'JD2', 'id': record_id, 'job_id': job_id},
-                                    })
-                                    logger.info('[JD2-QUEUE] Submitted %s for post-processing (folder: %s).', job_filename, nzb_folder)
-                                except Exception as err:
-                                    logger.warn('[JD2-QUEUE] Unable to enqueue %s for post-processing: %s', job_filename, err)
-                            else:
-                                logger.warn('[JD2-QUEUE] JD2 destination is not configured; skipping post-processing enqueue for %s.', job_filename)
-                        else:
-                            logger.info('[JD2-QUEUE] Post-processing disabled, please manually handle your files.')
-                        continue
-
-                    if job_status in failed_states:
-                        myDB.upsert(
-                                'ddl_info',
-                                {
-                                    'status': 'Failed',
-                                    'updated_date': datetime.datetime.now().strftime('%Y-%m-%d %H:%M'),
-                                },
-                                {'id': record_id},
-                            )
-                        logger.warn('[JD2-QUEUE] Download %s reported failure state (%s).', job_filename, job_status)
-                        continue
-
-                    if job_status not in completed_states:
-                        time.sleep(10)
-                        queue.put(item)
-                        continue
-            else:
-                logger.warn('[JD2-QUEUE] Missing JD2 client or job id for item: %s', item)
-        else:
-            time.sleep(10)
-
-
-def postprocess_main(queue):
-    while True:
-        if mylar.APILOCK is True:
-            time.sleep(5)
-
-        elif mylar.APILOCK is False and queue.qsize() >= 1: #len(queue) > 1:
-            pp = None
-            item = queue.get(True)
-            logger.info('Now loading from post-processing queue: %s' % item)
-            if item == 'exit':
-                logger.info('Cleaning up workers for shutdown')
-                break
-
-            if mylar.APILOCK is False:
-                try:
-                    pprocess = process.Process(item['nzb_name'], item['nzb_folder'], item['failed'], item['issueid'], item['comicid'], item['apicall'], item['ddl'], item['download_info'])
-                except:
-                    pprocess = process.Process(item['nzb_name'], item['nzb_folder'], item['failed'], item['issueid'], item['comicid'], item['apicall'])
-                pp = pprocess.post_process()
-                time.sleep(5) #arbitrary sleep to let the process attempt to finish pp'ing
-
-            if pp is not None:
-                if pp['mode'] == 'stop':
-                    #reset the lock so any subsequent items can pp and not keep the queue locked up.
-                    mylar.APILOCK = False
-
-            if mylar.APILOCK is True:
-                logger.info('Another item is post-processing still...')
-                time.sleep(15)
-                #mylar.PP_QUEUE.put(item)
-        else:
-            time.sleep(5)
-
-def search_queue(queue):
-    while True:
-        if mylar.SEARCHLOCK is True:
-            time.sleep(5)
-
-        elif mylar.SEARCHLOCK is False and queue.qsize() >= 1: #len(queue) > 1:
-            item = queue.get(True)
-            if item == 'exit':
-                logger.info('[SEARCH-QUEUE] Cleaning up workers for shutdown')
-                break
-
-            gumbo_line = True
-            #logger.fdebug('pack_issueids_dont_queue: %s' % mylar.PACK_ISSUEIDS_DONT_QUEUE)
-            #logger.fdebug('ddl_queued: %s' % mylar.DDL_QUEUED)
-            if item['issueid'] in mylar.PACK_ISSUEIDS_DONT_QUEUE:
-                if mylar.PACK_ISSUEIDS_DONT_QUEUE[item['issueid']] in mylar.DDL_QUEUED:
-                    logger.fdebug('[SEARCH-QUEUE-PACK-DETECTION] %s already queued to download via pack...Ignoring' % item['issueid'])
-                    gumbo_line = False
-
-            if gumbo_line:
-                logger.fdebug('[SEARCH-QUEUE] Now loading item from search queue: %s' % item)
-                if mylar.SEARCHLOCK is False:
-                    arcid = None
-                    comicid = item['comicid']
-                    issueid = item['issueid']
-                    if issueid is not None:
-                        if '_' in issueid:
-                            arcid = issueid
-                            comicid = None # required for storyarcs to work
-                            issueid = None # required for storyarcs to work
-                    mofo = mylar.filers.FileHandlers(ComicID=comicid, IssueID=issueid, arcID=arcid)
-                    local_check = mofo.walk_the_walk()
-
-                    if local_check['status']:
-                        fullpath = Path(local_check['filepath']) / local_check['filename']
-                        filecondition = check_file_condition(fullpath)
-                        if not filecondition['status']:
-                            logger.warn(f"CRC Check: File {fullpath} failed condition check ({filecondition['quality']}).  Ignoring.")
-                            local_check['status'] = False
-
-                    if local_check['status'] is True:
-                        mylar.PP_QUEUE.put({'nzb_name':     local_check['filename'],
-                                            'nzb_folder':   local_check['filepath'],
-                                            'failed':       False,
-                                            'issueid':      item['issueid'],
-                                            'comicid':      item['comicid'],
-                                            'apicall':      True,
-                                            'ddl':          False,
-                                            'download_info': None})
-                    else:
-                        try:
-                            manual = item['manual']
-                        except Exception as e:
-                            manual = False
-                        ss_queue = mylar.search.searchforissue(item['issueid'], manual=manual)
-                    time.sleep(5) #arbitrary sleep to let the process attempt to finish pp'ing
-
-            if mylar.SEARCHLOCK is True:
-                logger.fdebug('[SEARCH-QUEUE] Another item is currently being searched....')
-                time.sleep(15)
-        else:
-            time.sleep(5)
-
-
-def worker_main(queue):
-    while True:
-        if queue.qsize() >= 1:
-            item = queue.get(True)
-            logger.info('Now loading from queue: %s' % item)
-            if item == 'exit':
-                logger.info('Cleaning up workers for shutdown')
-                break
-            snstat = torrentinfo(torrent_hash=item['hash'], download=True)
-            if snstat['snatch_status'] == 'IN PROGRESS':
-                logger.info('Still downloading in client....let us try again momentarily.')
-                time.sleep(30)
-                mylar.SNATCHED_QUEUE.put(item)
-            elif any([snstat['snatch_status'] == 'MONITOR FAIL', snstat['snatch_status'] == 'MONITOR COMPLETE']):
-                logger.info('File copied for post-processing - submitting as a direct pp.')
-                mylar.PP_QUEUE.put({'nzb_name':     os.path.basename(snstat['copied_filepath']),
-                                    'nzb_folder':   snstat['copied_filepath'], #os.path.abspath(os.path.join(snstat['copied_filepath'], os.pardir)),
-                                    'failed':       False,
-                                    'issueid':      item['issueid'],
-                                    'comicid':      item['comicid'],
-                                    'apicall':      True,
-                                    'ddl':          False,
-                                    'download_info': None})
-                #threading.Thread(target=self.checkFolder, args=[os.path.abspath(os.path.join(snstat['copied_filepath'], os.pardir))]).start()
-        else:
-            time.sleep(15)
-
-def nzb_monitor(queue):
-    while True:
-        if mylar.RETURN_THE_NZBQUEUE.qsize() >= 1:
-            if mylar.USE_SABNZBD is True:
-                # this checks the sabnzbd queue to see if it's paused / unpaused.
-                sab_params = {
-                    'apikey': mylar.CONFIG.SAB_APIKEY,
-                    'mode': 'queue',
-                    'start': 0,
-                    'limit': 5,
-                    'search': None,
-                    'output': 'json',
-                }
-                s = sabnzbd.SABnzbd(params=sab_params)
-                sabresponse = s.sender(chkstatus=True)
-                # response will be: Paused = True, UnPaused = False
-                if sabresponse['status'] is False:
-                    while True:
-                        if mylar.RETURN_THE_NZBQUEUE.qsize() >= 1:
-                            qu_retrieve = mylar.RETURN_THE_NZBQUEUE.get(True)
-                            try:
-                                nzstat = s.historycheck(qu_retrieve)
-                                cdh_monitor(queue, qu_retrieve, nzstat, readd=True)
-                            except Exception as e:
-                                logger.error('Exception occured trying to re-add %s to queue: %s' % (qu_retrieve, e))
-                            time.sleep(5)
-                        else:
-                            break
-
-        if queue.qsize() >= 1:
-            item = queue.get(True)
-            if item == 'exit':
-                logger.info('Cleaning up workers for shutdown')
-                break
-            try:
-                tmp_apikey = item['queue'].pop('apikey')
-                logger.info('Now loading from queue: %s' % item)
-            except Exception:
-                #nzbget doesn't pass the queue field. So just let it fly.
-                logger.info('Now loading from queue: %s' % item)
-            else:
-                item['queue']['apikey'] = tmp_apikey
-            if all([mylar.USE_SABNZBD is True, mylar.CONFIG.SAB_CLIENT_POST_PROCESSING is True]):
-                nz = sabnzbd.SABnzbd(item)
-                nzstat = nz.processor()
-            elif all([mylar.USE_NZBGET is True, mylar.CONFIG.NZBGET_CLIENT_POST_PROCESSING is True]):
-                nz = nzbget.NZBGet()
-                nzstat = nz.processor(item)
-            else:
-                logger.warn('There are no NZB Completed Download handlers enabled. Not sending item to completed download handling...')
-                break
-            cdh_monitor(queue, item, nzstat)
-        else:
-            time.sleep(5)
-
-
-def cdh_monitor(queue, item, nzstat, readd=False):
-    known_nzb_id = item['nzo_id'] if (mylar.USE_SABNZBD is True) else item['NZBID']
-    if any([nzstat['status'] == 'file not found', nzstat['status'] == 'double-pp']):
-        logger.warn('Unable to complete post-processing call due to not finding file in the location provided. [%s]' % item)
-    elif nzstat['status'] == 'nzb removed' or 'unhandled status' in str(nzstat['status']).lower():
-        if readd is True:
-            # if the queue isn't empty, retry it like 5 times or until queue is empty
-            # if queue is empty, this could be a timing issue, so we'd have to recheck it one more time at least
-            # need to implement some kind of counter here for all of this ^^^
-            logger.warn('NZB seems to have been in a staging process within SABnzbd during attempt. Will requeue: %s.' % known_nzb_id)
-            mylar.RETURN_THE_NZBQUEUE.put(item)
-        else:
-            logger.warn('NZB seems to have been removed from queue: %s' % known_nzb_id)
-    elif nzstat['status'] == 'failed_in_sab':
-        logger.warn('Failure returned from SAB for %s for some reason. You should probably check your SABnzbd logs' % known_nzb_id)
-    elif nzstat['status'] == 'queue_paused':
-        # queue pause check for sabnzbd only atm.
-        if mylar.USE_SABNZBD is True:
-            logger.info('[PAUSED_SAB_QUEUE] adding %s to a temporary queue that will fire off when SABnzbd is unpaused' % item)
-            mylar.RETURN_THE_NZBQUEUE.put(item)
-    elif nzstat['status'] is False:
-        logger.info('Download %s failed. Requeue NZB to check later...' % known_nzb_id)
-        time.sleep(5)
-        if item not in queue.queue:
-            mylar.NZB_QUEUE.put(item)
-    elif nzstat['status'] is True:
-        # Currently only checks SAB at this point as nzbget filename not available
-        if nzstat['failed'] is False and mylar.USE_SABNZBD is True:
-            fullpath = Path(nzstat['location']) / nzstat['name']
-            filecondition = check_file_condition(fullpath)
-            if not filecondition['status']:
-                logger.warn(f"CRC Check: File {fullpath} failed condition check ({filecondition['quality']}).  Marking as failed.")
-                nzstat['failed'] = True
-
-        if nzstat['failed'] is False:
-            logger.info('File successfully downloaded - now initiating completed downloading handling.')
-        else:
-            logger.info('File failed either due to being corrupt or incomplete - now initiating completed failed downloading handling.')
-        try:
-            mylar.PP_QUEUE.put({'nzb_name':     nzstat['name'],
-                                'nzb_folder':   nzstat['location'],
-                                'failed':       nzstat['failed'],
-                                'issueid':      nzstat['issueid'],
-                                'comicid':      nzstat['comicid'],
-                                'apicall':      nzstat['apicall'],
-                                'ddl':          False,
-                                'download_info': nzstat['download_info']})
-        except Exception as e:
-            logger.error('process error: %s' % e)
-    return
-
-
-QueueInfo = namedtuple("QueueInfo", ("name", "is_alive", "size"))
-
-
-def queue_info():
-    yield from (
-        QueueInfo(queue_name, thread_obj.is_alive() if thread_obj is not None else None, queue.qsize())
-        for (queue_name, thread_obj, queue) in [
-            ("AUTO-COMPLETE-NZB", mylar.NZBPOOL, mylar.NZB_QUEUE),
-            ("AUTO-SNATCHER", mylar.SNPOOL, mylar.SNATCHED_QUEUE),
-            ("DDL-QUEUE", mylar.DDLPOOL, mylar.DDL_QUEUE),
-            ("JD2-QUEUE", mylar.JD2POOL, mylar.JD2_QUEUE),
-            ("POST-PROCESS-QUEUE", mylar.PPPOOL, mylar.PP_QUEUE),
-            ("SEARCH-QUEUE", mylar.SEARCHPOOL, mylar.SEARCH_QUEUE),
-        ]
-    )
-
 
 def script_env(mode, vars):
     #mode = on-snatch, pre-postprocess, post-postprocess
